pallet_asset_rewards::pallet

Trait Config

pub trait Config: Config {
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type PalletId: Get<PalletId>;
    type AssetId: AssetId + Member + Parameter;
    type Balance: Balance + TypeInfo;
    type CreatePoolOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>;
    type Assets: Inspect<Self::AccountId, AssetId = Self::AssetId, Balance = Self::Balance> + Mutate<Self::AccountId>;
    type AssetsFreezer: MutateFreeze<Self::AccountId, Id = Self::RuntimeFreezeReason, AssetId = Self::AssetId, Balance = Self::Balance>;
    type RuntimeFreezeReason: From<FreezeReason>;
    type Consideration: Consideration<Self::AccountId, Footprint>;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

The main purpose of this trait is to act as an interface between this pallet and the runtime in which it is embedded in. A type, function, or constant in this trait is essentially left to be configured by the runtime that includes this pallet.

Consequently, a runtime that wants to include this pallet must implement this trait.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Overarching event type.

type PalletId: Get<PalletId>

The pallet’s unique identifier, used to derive the pool’s account ID.

The account ID is derived once during pool creation and stored in the storage.

type AssetId: AssetId + Member + Parameter

Identifier for each type of asset.

type Balance: Balance + TypeInfo

The type in which the assets are measured.

type CreatePoolOrigin: EnsureOrigin<Self::RuntimeOrigin, Success = Self::AccountId>

The origin with permission to create pools.

The Origin must return an AccountId.

type Assets: Inspect<Self::AccountId, AssetId = Self::AssetId, Balance = Self::Balance> + Mutate<Self::AccountId>

Registry of assets that can be configured to either stake for rewards, or be offered as rewards for staking.

type AssetsFreezer: MutateFreeze<Self::AccountId, Id = Self::RuntimeFreezeReason, AssetId = Self::AssetId, Balance = Self::Balance>

Freezer for the Assets.

type RuntimeFreezeReason: From<FreezeReason>

The overarching freeze reason.

type Consideration: Consideration<Self::AccountId, Footprint>

Means for associating a cost with the on-chain storage of pool information, which is incurred by the pool creator.

The passed Footprint specifically accounts for the storage footprint of the pool’s information itself, excluding any potential storage footprint related to the stakers.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.